+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 master
 ------
 
+* Prepend Sprockets path to EmberCLI assets.
+  Default fingerprinting to no generate `MD5` hash when building in
+  `development`.  [#17].
+
+[#17]: https://github.com/rondale-sc/ember-cli-rails-addon/pull/17
+
 0.0.13
 ------